+++ pennylane/ops/qubit/state_preparation.py
@@ -17,14 +17,9 @@ with preparing a certain state on the device.
 """
 # pylint:disable=abstract-method,arguments-differ,protected-access
 
+import pennylane as qml
 from pennylane.operation import AnyWires, Operation
 from pennylane.templates.state_preparations import BasisStatePreparation, MottonenStatePreparation
-
-
-class AdjointError(Exception):
-    """Exception for non-adjointable operations."""
-
-    pass
 
 
 class BasisState(Operation):
@@ -59,7 +54,7 @@ class BasisState(Operation):
         return BasisStatePreparation(n, wires)
 
     def adjoint(self):
-        raise AdjointError("No adjoint exists for BasisState operations.")
+        raise qml.ops.AdjointError("No adjoint exists for BasisState operations.")
 
 
 class QubitStateVector(Operation):
